+++ pkgs/ffigen/test/large_integration_tests/large_objc_test.dart
@@ -111,8 +111,8 @@ void main() {
           ),
           global: (node) =>
               node.isIncluded = stableRandomInclude('globals', node),
-          typealias: (node) =>
-              node.isIncluded = stableRandomInclude('typedefs', node),
+          typealias: (node) => node.isIncluded =
+              stableRandomInclude('typedefs', node) ? .ifUsed : .never,
           macroConstant: (node) => node.isIncluded = false,
           objCInterface: (node) {
             node.includeCategories = false;