+++ pkgs/swiftgen/lib/src/generator.dart
@@ -86,7 +86,6 @@ extension SwiftGenGenerator on SwiftGenerator {
         style: fg.NativeExternalBindings(assetId: output.assetId),
       ),
       functions: ffigen.functions,
-      typedefs: ffigen.typedefs,
       objectiveC: fg.ObjectiveC(
         externalVersions: ffigen.objectiveC.externalVersions,
       ),